backend/pyproject.toml
[project]
name = "insurance-dashboard-api"
version = "1.0.0"
description = "FastAPI backend for Argentine insurance market dashboard"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pandas>=2.1.3",
    "pyarrow>=14.0.1",
    "python-dotenv>=1.0.0",
]
